+++ src/components/form/select.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="mdl-select">
+  </div>
+</template>
+
+<script>
+// import UiTextfield from './textfield';
+// import UiMenu from '../menu';
+
+// export default {
+//   name: 'ui-select'
+// };
+</script>