+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 ## 0.1.2 beta
 
-## Update components
+### Update components
 
 __rename `<ui-splitter>` to `<ui-divider>`__
+
+__`<ui-table>`__
+
+- fix checkbox event
+- rename props `checkList` to `checkboxList`
+- add tbody attributes
+- add no data show
+
+### Update docs
+
+- `<ui-radio>`
+- `<ui-table>`
 
 ## 0.1.1 beta